Quit throwing exceptions on output on STDERR channel (fixes #61)
PHP has started writing its PHP Errors to STDERR in CLI which means an expected
error within the test case could cause PHPT to mis-interpret the error as an
errored test.  Instead of making the assumption that no STDERR output should be
generated, PHPT now adds it to the output.

BC BREAK: This breaks existing behavior which expected all PHP errors to be
displayed as part of STDOUT and anything on STDERR was automatically considered
bad.


git-svn-id: https://svn.domain51.com/test/Core/trunk@695 54d25d7b-d40e-47af-a99d-8d4037bc02b6