Fixes false failures caused by test cases with Windows EOL markers (re: #79)
The visibility_005.phpt.inc file was taken directly from php-src, where it
was first observed to fail with PHPT.  The problem was the \r\n EOL markers.
This change strips the EOL markers off of each line of a case that it parses
and converts them into \n so they are standardized.

This may also help issues where diffs are not generated properly.